refactor: abstract graphql mutations
